# project2-api-pokemon/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: API Pokemon
  description: An API to manage Pokemon
host: localhost:8080
basePath: /api
schemes:
  - http
paths:
  /pokemons:
    get:
      description: get the list of all pokemons
      operationId: getPokemons
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Pokemon'

    delete:
      operationId: deletePokemons
      description: delete the list of all pokemons
      responses:
        204:
          description: sucessfully deleted

    post:
      description: create a pokemon
      operationId: createPokemon
      consumes:
        - application/json
      parameters:
        - name: pokemon
          in: body
          required: true
          schema:
            $ref: '#/definitions/Pokemon'
      responses:
        '201':
          description: created
          schema:
            $ref: '#/definitions/Pokemon'
        '400':
          description: Bad request

  /pokemons/{id}:
    get:
      description: get a pokemon by its ID
      operationId: getPokemonByID
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Pokemon'
        '404':
          description: pokemon not found
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The pokemon ID

    delete:
      description: delete a pokemon by its ID
      operationId: deletePokemonByID
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The pokemon ID
      responses:
        204:
          description: sucessfully deleted
        '404':
          description: Not found

    put:
      description: update a pokemon by its ID
      operationId: updatePokemonByID
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The pokemon ID
        - name: pokemon
          in: body
          required: true
          schema:
            $ref: '#/definitions/Pokemon'
      responses:
        200:
          description: success
        '404':
          description: Not found

  /trainers:
    get:
      description: get the list of all trainers
      operationId: getTrainers
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/TrainerWithId'

    delete:
      operationId: deleteTrainers
      description: delete the list of all trainers
      responses:
        204:
          description: sucessfully deleted

    post:
      description: create a trainer
      operationId: createTrainer
      consumes:
        - application/json
      parameters:
        - name: trainer
          in: body
          required: true
          schema:
            $ref: '#/definitions/Trainer'
      responses:
        '201':
          description: created
          schema:
            $ref: '#/definitions/TrainerWithId'
        '400':
          description: Bad request

  /trainers/{id}:
    get:
      description: get a trainer by its ID
      operationId: getTrainerById
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/TrainerWithId'
        '404':
          description: Not found
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The trainer ID

    delete:
      description: delete a trainer by its ID
      operationId: deleteTrainerById
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The trainer ID
      responses:
        204:
          description: sucessfully deleted
        '404':
          description: Not found

    put:
      description: update a trainer by its ID
      operationId: updateTrainerById
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The trainer ID
        - name: trainer
          in: body
          required: true
          schema:
            $ref: '#/definitions/Trainer'
      responses:
        200:
          description: success
        '404':
          description: Not found

  /captures:
    post:
      description: create a capture
      operationId: createCapture
      consumes:
        - application/json
      parameters:
        - name: capture
          in: body
          required: true
          schema:
            $ref: '#/definitions/Capture'
      responses:
        '201':
          description: created
          schema:
            $ref: '#/definitions/CaptureWithId'
        '400':
          description: Bad request

  /captures/trainers/{id_trainer}:
    get:
      description: Get a specific trainer with his all pokemons
      operationId: getTrainerWithPokemons
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/CaptureGetAll'
        '404':
          description: Not found
      parameters:
        - in: path
          name: id_trainer
          required: true
          type: integer
          description: trainer ID

  /captures/pokemons/{id_pokemon}:
    get:
      description : Get a specific pokemon with his all trainers
      operationId: getPokemonWithTrainers
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/CaptureGetAll'
        '404':
          description: Not found
      parameters:
        - in: path
          name: id_pokemon
          required: true
          type: integer
          description: pokemon ID

definitions:
  Pokemon:
    type: object
    properties:
      pokedexId:
        type: integer
      idUser:
        type: integer
      name:
        type: string
      type:
        type: string
      category:
        type: string
      height:
        type: integer
      hp:
        type: integer

  Trainer:
    type: object
    properties:
      idUser:
        type: integer
      name:
        type: string
      surname:
        type: string
      gender:
        type: string
      age:
        type: integer
      numberOfBadges:
        type: integer

  TrainerWithId:
    allOf:
      - type: object
        required:
          - id
        properties:
          id:
            type: integer
      - $ref: '#/definitions/Trainer'

  Capture:
    type: object
    properties:
      idUser:
        type: integer
      idTrainer:
        type: integer
      pokemons:
        type: array
        items:
          $ref: '#/definitions/CapturePokemon'

  CaptureWithId:
    allOf:
      - type: object
        properties:
          id:
            type: integer
      - $ref: '#/definitions/Capture'

  CapturePokemon:
    type: object
    properties:
      idPokemon:
        type: integer
      dateCapture:
        type: string

  CaptureGet:
    type: object
    properties:
      idUser:
        type: integer
      idCapture:
        type: integer
      idPokemon:
        type: integer
      idTrainer:
        type: integer
      dateCapture:
        type: string

  CaptureGetAll:
    type: object
    properties:
      captures:
        type: array
        items:
          $ref: '#/definitions/CaptureGet'